vp - Rename files to match nteract pattern
