+++ packages/jupyter-widgets/src/manager/widget-loader.ts
